requirements.txt
alembic==1.6.5
aniso8601==8.0.0
bcrypt==3.1.7
bidict==0.21.2
blinker==1.4
certifi==2021.5.30
cffi==1.14.0
charset-normalizer==2.0.3
click==7.1.1
cryptography==3.4.7
decorator==5.0.9
dictdiffer==0.8.1
dnspython==1.16.0
eventlet==0.31.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.0.4
idna==3.2
importlib-metadata==4.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpath-rw==1.4.0
Mako==1.1.4
MarkupSafe==1.1.1
migrate==0.3.8
mongoengine==0.23.1
mongoengine-migrate==0.0.2
numpy==1.19.5
phonenumbers==8.12.27
ply==3.11
psycopg2-binary==2.9.1
pycparser==2.20
pycrypto==2.6.1
PyJWT==2.1.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2019.3
requests==2.26.0
six==1.14.0
SQLAlchemy==1.4.21
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.2.1
zipp==3.5.0
